+++ crates/firstlight-view/src/app.rs
@@ -116,6 +116,7 @@ pub struct FirstLightApp {
     display_times: VecDeque<Instant>,
 
     pub auto_stretch: bool,
+    pub neutralise_colour: bool,
     pub gamma: f32,
     pub debayer: bool,
     pub display: DisplayOptions,
@@ -157,6 +158,7 @@ impl FirstLightApp {
             last_levels: (0, 0),
             display_times: VecDeque::new(),
             auto_stretch: true,
+            neutralise_colour: true,
             gamma: 1.0,
             debayer: true,
             display: DisplayOptions::default(),
@@ -359,6 +361,7 @@ impl FirstLightApp {
             } else {
                 Stretch::Linear
             },
+            neutralise_colour: self.neutralise_colour,
             debayer: self.debayer,
             subsample,
             gamma: self.gamma,